+++ MekHQ/src/mekhq/campaign/mission/newContract/AbstractContract.java
@@ -53,7 +53,6 @@ import mekhq.campaign.mission.enums.ContractObjectiveType;
 import mekhq.campaign.mission.enums.MissionStatus;
 import mekhq.campaign.mission.newContract.contractData.*;
 import mekhq.campaign.mission.newContract.contractGeneration.ChaosEmployerType;
-import mekhq.campaign.mission.newContract.contractGeneration.ContractTermsData;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.universe.Faction;
 import mekhq.campaign.universe.Planet;
