+++ MekHQ/src/mekhq/campaign/mission/newContract/contractGeneration/ChaosContractDetermineTerms.java
@@ -35,6 +35,7 @@ package mekhq.campaign.mission.newContract.contractGeneration;
 import static megamek.common.compute.Compute.d6;
 
 import mekhq.campaign.mission.newContract.contractData.ChaosContractStepsTable;
+import mekhq.campaign.mission.newContract.contractData.ContractTermsData;
 
 public class ChaosContractDetermineTerms {
     public static ContractTermsData determineInitialTerms(ChaosObjectiveType objectiveType,
