+++ MekHQ/src/mekhq/gui/dialog/markets/contractMarket/ContractNegotiationDialog.java
@@ -69,9 +69,9 @@ import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.finances.Money;
 import mekhq.campaign.mission.newContract.AbstractContract;
 import mekhq.campaign.mission.newContract.contractData.ChaosContractStepsTable;
+import mekhq.campaign.mission.newContract.contractData.ContractTermsData;
 import mekhq.campaign.mission.newContract.contractData.NegotiationData;
 import mekhq.campaign.mission.newContract.contractData.RentedFacilitiesData;
-import mekhq.campaign.mission.newContract.contractGeneration.ContractTermsData;
 import mekhq.campaign.mission.newContract.contractGeneration.negotiationsAndNPCs.TermFunding;
 import mekhq.campaign.mission.newContract.contractGeneration.targetFinder.ChaosContractPayDetermination;
 import mekhq.campaign.personnel.Person;
